+++ src/Models/ScriptConfig.php
@@ -19,12 +19,26 @@ class ScriptConfig extends BaseServiceConfigModel
 {
     protected $table = 'script_config';
 
-    protected $fillable = ['service_id', 'content', 'config', 'queued'];
+    protected $fillable = [
+        'service_id',
+        'content',
+        'config',
+        'queued',
+        'storage_service_id',
+        'scm_reference',
+        'scm_repository',
+        'storage_path'
+    ];
 
     // deprecated, service has type designation now
     protected $hidden = ['type'];
 
-    protected $casts = ['service_id' => 'integer', 'config' => 'array', 'queued' => 'boolean'];
+    protected $casts = [
+        'service_id'         => 'integer',
+        'config'             => 'array',
+        'queued'             => 'boolean',
+        'storage_service_id' => 'integer',
+    ];
 
     /**
      * @var array Extra config to pass to any config handler
@@ -95,6 +109,9 @@ class ScriptConfig extends BaseServiceConfigModel
                     '- QUEUED_DELAY = #seconds to delay the execution of the script.<br>' .
                     '- QUEUED_QUEUE= alternative queue from the system configuration.<br>' .
                     '- QUEUED_CONNECTION = alternative queue connection from the system configuration.';
+                break;
+            case 'storage_service_id':
+                $schema['type'] = 'integer';
                 break;
         }
     }